+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Improved scene switch performance (thanks @Phantomical )
 - Fixed NRE spam when audio sources are destroyed in certain ways
+- Fixed NRE spam when a RSE_RCS module was added to a part that doesn't have RCS
 
 
 ## [0.9.11] - 03-25-24